backend/pyproject.toml
[project]
name = "finalerts"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.14.0",
    "fastapi[standard]>=0.115.4",
    "httpx>=0.27.2",
    "jose>=1.0.0",
    "pandas>=2.2.3",
    "passlib>=1.7.4",
    "pydantic-settings>=2.6.1",
    "pyjwt>=2.10.0",
    "requests>=2.32.3",
    "sqlalchemy>=2.0.36",
    "uvicorn>=0.32.0",
]
